VIM-3001 move motions with textobj-user
